Add blacklist options
This adds a new configuration item: `DNSBL`.
This is checked against non-logged in users on every page. If they are not logged in they get blacklist checked and then redirected to the login page, providing their IP was not blacklisted.

If it was blacklisted the user will be redirected to an error screen with a cute dog on it.

I have also spent a considerable amount of time trying (and failing) to make it look good on mobile devices, so this is a small commit ;)