Rather large update, please see commit notes
With this update:
- WebPanel version shows in the footer for logged-in users (sql_auth)
- Made things look much better on mobiles; before, the navigation bar covered most of the mobile screen and was generally not visually appealing. So now when the screen is a mobile, the navigation bar will collapse into a collapsible menu.
- Separated `Name Bans` and `Server Ban Exceptions` into their own sub-menu pages. TODO: There is some weird overlap with the list background highlight.
-Started on user_meta: as in, when they last logged in, what IP they are using or used, recent actions taken, and most importantly, permissions.
- Added a few hooks which allow plugins to do more:
  - `HOOKTYPE_PRE_HEADER`
  - `HOOKTYPE_HEADER`
  - `HOOKTYPE_PRE_FOOTER`
  - `HOOKTYPE_FOOTER`

These hooks are relatively self-explanatory. There is however another hook of a different nature.

- `HOOKTYPE_NOTIFICATION`
You will need to run this from your plugin as well as catch functions from it. This won't be run many places in the source so it's up to you to send notifications. Maybe some other plugin wants to show notifications somewhere else or do something different, while still being able to load your plugin and receive notifications for it, and vice versa.