Update unrealircd-admin.css
Tested on Firefox: as soon as we are on any page, there is the scrollbar activated at the bottom of page.
The div that could pose a problem is the div that contains the nav tag with the "container-fluid" class, the default bootstrap contains a width: 100% and this width when it is removed the scrollbar is removed.
This css disables/resets the width. It seems to work pretty well on my Firefox browser.