Add the start of "Network Health" tool
This tab will let you see if there are any mode mismatches or module version mismatches, and any other stuff we can think might be useful to add in.
Since we already have a hardcoded End Of Life, it sounds reasonable to add UnrealIRCd5 being EOL as a  "problem" listed on this page after July, for example.