Creation of a backup folder do as anope_user
Will this PHP file be useful in the future?
Currently, it saves the list of logged-in users directly into an SQL table named unreal_irc_users. The SQL table is created automatically if it does not exist.
This PHP file was created to be run by cron every 1 minute, but it is unfortunate that it is not updated in real-time. To achieve real-time updates, a Node.JS bot that updates each connection and disconnection may be necessary.