Minimize config/config.php and put settings in DB. This is work in progress.
config/config.php is no longer meant to be edited by hand and will
only contain the bare minimum of config that needs to be in a file.
All the rest will be in the DB (either SQL or db file).

Various todo items and actual db-based settings read/write untested.