Some work on right-click menus and other things
This is largely unavailable yet because it's unfinished, for example the right-click menu in firefox can't figure out copy and paste, which is annoying.

Working on a feedback dialog which pops up and asks for feedback every 3 months (unifinished, but that's the plan).

The right-click menu will aim to be hookable in  PHP as well as javascript, i.e. you can add to it via javascript within your plugin or you could run a hook so your option displays on the api endpoint, which is probably a bit cleaner.