updated to work on apache2
The lists of users and chat rooms... they are not functioning because users.php, for example, does not execute due to it stopping prematurely because of common_api.php. Disabling this, it works fine with Apache2. The same identical issue occurs on two different servers with Apache2: one before December 2023 on CentOS with Cpanel, and the same on Debian 12 with Apache2 as well.