Allow installing as a mobile app (Progressive Web App)
This allows you to install your webpanel to your phone as a native application using an embedded browser (Progressive Web App) to look and function like a normal app on your mobile device. Just go to "Save this page" and it should give you the option to install as an app on your phone or tablet.

If you are installing this fresh, it should just work immediately.

If you are upgrading your webpanel, you need to re-login for the PWA functionality to kick in.